+++ Sources/SpotifyWebAPI/Object Model/Post Request Objects/Playlist Objects/URIsDictWithInsertionIndex.swift
@@ -1,9 +1,11 @@
 import Foundation
 
 /**
- Contains an array of URIs and (optionally) the position
- to insert them in the playlist. Used in the body of `addToPlaylist`.
+ Contains an array of URIs and (optionally) the position to insert them
+ in a playlist. Used in the body of
+ `SpotifyAPI.addToPlaylist(_:uris:position:)`.
  
+ For example:
  ```
  {
     "uris": [
@@ -17,15 +19,81 @@ import Foundation
 
  [1]: https://developer.spotify.com/documentation/web-api/reference/playlists/add-tracks-to-playlist/
  */
-struct URIsDictWithInsertionIndex: Codable, Hashable {
+public struct URIsDictWithInsertionIndex {
     
-    let uris: [String]
-    let position: Int?
+    /// An array of track/episode URIs that will be added to a playlist.
+    public var uris: [SpotifyURIConvertible]
     
-    init(
-        uris: [SpotifyURIConvertible], postion: Int?
+    /// The zero-indexed position at which to insert `uris` in a playlist.
+    /// If `nil`, then the `uris` will be appended to the playlist.
+    public var position: Int?
+    
+    /**
+     Creates an array of URIs and the position to insert them in a playlist.
+     
+     - Parameters:
+       - uris: An array of URIs.
+       - position: The position to insert the URIs in a playlist. If `nil`
+             (default), then they will be appended to the playlist.
+     */
+    public init(
+        uris: [SpotifyURIConvertible], position: Int? = nil
     ) {
-        self.uris = uris.map(\.uri)
-        self.position = postion
+        self.uris = uris
+        self.position = position
     }
+    
+}
+
+extension URIsDictWithInsertionIndex: Codable {
+
+    /// :nodoc:
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        
+        self.uris = try container.decode(
+            [String].self, forKey: .uris
+        )
+        
+        self.position = try container.decodeIfPresent(
+            Int.self, forKey: .position
+        )
+        
+    }
+    
+    /// :nodoc:
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        
+        try container.encode(
+            self.uris.map(\.uri), forKey: .uris
+        )
+        try container.encodeIfPresent(
+            self.position, forKey: .position
+        )
+    }
+
+    /// :nodoc:
+    public enum CodingKeys: String, CodingKey {
+        case uris
+        case position
+    }
+
+}
+
+
+extension URIsDictWithInsertionIndex: Hashable {
+    
+    /// :nodoc:
+    public func hash(into hasher: inout Hasher) {
+        hasher.combine(position)
+        hasher.combine(uris.map(\.uri))
+    }
+    
+    /// :nodoc:
+    public static func == (lhs: Self, rhs: Self) -> Bool {
+        return lhs.position == rhs.position &&
+                lhs.uris.map(\.uri) == rhs.uris.map(\.uri)
+    }
+
 }