+++ CHANGELOG.md
@@ -4,7 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.2] - 2020-11-2
+## [1.0.3] - 2020-11-09
+
+### Added
+
+* Added `URIsWithPositionsContainer.init(snapshotId:urisWithSinglePosition:)` and `URIsWithPositionsContainer.chunked(urisWithSinglePosition:)`. These methods will aid in removing more than 100 duplicate items from a playlist.
+
+### Changed
+
+* `URIsDictWithInsertionIndex` is now public
+* `URIsContainer.init(_:snapshotId:)` and `URIsWithPositionsContainer.init(snapshotId:urisWithPositions:)` now have a default value of `nil` for the snapshot id.
+
+### Deprecated
+
+* Deprecated `URIsWithPositionsContainer` initializer that accepted an array of tuples.
+
+## [1.0.2] - 2020-11-02
 
 ### Changed